pyproject.toml
[project]
name = "hi"
version = "0.1.0"
description = "Healthcare Informatics Skills Framework CLI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.0",
    "ruamel.yaml>=0.18",
]

[project.scripts]
hi = "hi.cli:main"

[project.urls]
Repository = "https://github.com/your-org/hi-skills"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hi"]

[tool.hatch.build.targets.wheel.force-include]
"src/hi/schemas" = "hi/schemas"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.uv]
# End users install with: uv tool install hi
# Dev setup: uv sync && uv run hi --help

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
